Implemented binary filter and marker
Implemented a binary filter (black and white) to identify the colour yellow.

Implemented code to draw a marker at the centre of mass of the yellow colour.

played around with the setServo functions provided in the owl.h library.

very close to implementing tracking.

Implementing tracking will be the next action to be performed. Then I will tidy  up the code and make sure to CHECK FOR ZERO ERROR